let proposers using disjoint numbers decide among accepted vals
"Paxos Made Simple" makes it clear that the acceptors must retain
their previously accepted values when they see a "Prepare" message
with a higher number.  They will still Accept a value with a higher
number:

  "If an acceptor receives an accept request for a
   proposal numbered n, it accepts the proposal
   unless it has already responded to a prepare
   request having a number greater than n."

The proposer gets all the responses to the Prepare messages and
selects the value with the highest associated number from them.
Only when there are no returned accepted values can the proposer
choose some new value for the Accept message.